# Statistics update optimization