@benjaminwinger 2025-03-03
It looks like the performance discrepancy is because we're partitioning and it is appending a single value at a time. We should probably buffer the appended values, or delay the statistics updates somehow, as it would be much more efficient to calculate the statistics in batches, or even as a whole after the partitioner is done.